requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==4.0
attrs==23.1.0
billiard==4.1.0
cachetools==5.3.1
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
cryptography==41.0.1
dictdiffer==0.9.0
diskcache==5.6.1
distro==1.8.0
dpath==2.1.6
dulwich==0.21.5
dvc==3.0.0
dvc-data==1.11.0
dvc-gdrive==2.19.2
dvc-http==2.30.2
dvc-objects==0.23.0
dvc-render==0.5.3
dvc-studio-client==0.10.0
dvc-task==0.3.0
filelock==3.12.2
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.3.3
fsspec==2023.6.0
funcy==2.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
grandalf==0.8
httplib2==0.22.0
hydra-core==1.3.2
idna==3.4
iterative-telemetry==0.0.8
joblib==1.2.0
kombu==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
nanotime==0.5.2
networkx==3.1
numpy==1.24.3
oauth2client==4.1.3
omegaconf==2.3.0
orjson==3.9.1
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
platformdirs==3.5.3
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
PyDrive2==1.15.4
pygit2==1.12.1
Pygments==2.15.1
pygtrie==2.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0
requests==2.31.0
rich==13.4.2
rsa==4.9
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
scmrepo==1.0.3
shortuuid==1.0.11
shtab==1.6.2
six==1.16.0
smmap==5.0.0
sqltrie==0.6.0
tabulate==0.9.0
threadpoolctl==3.1.0
tomlkit==0.11.8
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
yarl==1.9.2
zc.lockfile==3.0.post1
